requirements.txt
Flask
python-dotenv
requests
base64
os
random